+++ src/Generics/Deriving/Semigroup/Internal.hs
@@ -189,9 +189,6 @@ instance Ord a => GSemigroup (Min a) where
 
 instance GSemigroup (NonEmpty a) where
   gsappend = (<>)
-
-instance GSemigroup a => GSemigroup (Option a) where
-  gsappend (Option a) (Option b) = Option (gsappend a b)
 #endif
 
 -- Tuple instances