+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # next [????.??.??]
+* Allow building with `template-haskell-2.17.0.0` (GHC 9.0).
 * Fix a bug in which `deriveAll1` would needlessly reject data types whose last
   type parameter appears as an oversaturated argument to a type family.